Split angle workspace helpers
