+++ frontend/src/components/tools/AngleWorkspace.tsx
@@ -25,275 +25,45 @@ import { HeaderBar } from '@/components/HeaderBar';
 import { AppSidebar } from '@/components/AppSidebar';
 import { Button, ButtonLink } from '@/components/ui/Button';
 import { Card } from '@/components/ui/Card';
-import { listAngleToolEngines } from '@/config/tools-angle-engines';
 import { runAngleTool, saveImageToLibrary, useInfiniteJobs } from '@/lib/api';
-import { dispatchAnalyticsEvent } from '@/lib/analytics-client';
 import { authFetch } from '@/lib/authFetch';
-import { prepareImageFileForUpload } from '@/lib/client-image-upload';
 import { suggestDownloadFilename, triggerAppDownload } from '@/lib/download';
 import { useI18n } from '@/lib/i18n/I18nProvider';
 import { resolveAngleEngineForParams } from '@/lib/tools-angle';
 import { useRequireAuth } from '@/hooks/useRequireAuth';
 import { FEATURES } from '@/content/feature-flags';
 import type { AngleToolEngineId, AngleToolNumericParams, AngleToolResponse } from '@/types/tools-angle';
-
-type UploadedImage = {
-  id?: string | null;
-  url: string;
-  previewUrl?: string | null;
-  width?: number | null;
-  height?: number | null;
-  name?: string | null;
-  source?: 'upload' | 'library' | 'paste' | 'example';
-};
-
-type LibraryAsset = {
-  id: string;
-  url: string;
-  mime?: string | null;
-  width?: number | null;
-  height?: number | null;
-  size?: number | null;
-  source?: string | null;
-  createdAt?: string;
-};
-
-type LibraryAssetsResponse = {
-  ok: boolean;
-  assets: LibraryAsset[];
-};
-
-type BillingProductResponse = {
-  ok: boolean;
-  product?: {
-    productKey: string;
-    currency: string;
-    unitPriceCents: number;
-  };
-  error?: string;
-};
-
-type AnglePreviewImage = {
-  url: string;
-  thumbUrl?: string | null;
-};
-
-type RecentAngleJobEntry = {
-  jobId: string;
-  createdAt: string;
-  engineLabel: string;
-  outputs: AnglePreviewImage[];
-};
+import { DEFAULT_ANGLE_COPY, type AngleCopy } from './angle/_lib/angle-workspace-copy';
+import {
+  ANGLE_GUEST_EXAMPLE_OUTPUT_URL,
+  ANGLE_GUEST_EXAMPLE_SOURCE_URL,
+  ANGLE_TOOL_STORAGE_KEY,
+  cleanupSourcePreview,
+  collectAnglePreviewImages,
+  DEFAULT_ENGINE_ID,
+  emitClientMetric,
+  ENGINES,
+  formatUsdCompact,
+  getAngleBillingProductKey,
+  isAuthRequiredError,
+  parsePersistedAngleToolState,
+  sanitizeParams,
+  uploadImage,
+} from './angle/_lib/angle-workspace-helpers';
+import type {
+  AnglePreviewImage,
+  BillingProductResponse,
+  LibraryAsset,
+  LibraryAssetsResponse,
+  PersistedAngleToolState,
+  RecentAngleJobEntry,
+  UploadedImage,
+} from './angle/_lib/angle-workspace-types';
 
 const AngleOrbitCanvas = dynamic(() => import('@/components/tools/AngleOrbitCanvas'), {
   ssr: false,
   loading: () => <div className="h-full w-full bg-[#f6f8fc]" aria-hidden="true" />,
 });
-
-const DEFAULT_ANGLE_COPY = {
-  "disabledTitle": "Tools are disabled",
-  "disabledBody": "Enable `FEATURES.workflows.toolsSection` to access this area.",
-  "back": "Back to Tools",
-  "engine": "Engine",
-  "engineSelect": "Engine selection",
-  "sourceImage": "Source image",
-  "sourceReady": "Source image ready",
-  "sourceFromLibrary": "From Library",
-  "sourceFromPaste": "From pasted URL",
-  "sourceFromDevice": "Uploaded from device",
-  "sourceFromExample": "Public example",
-  "replace": "Replace",
-  "library": "Library",
-  "remove": "Remove",
-  "sourceAlt": "Source",
-  "addSourceTitle": "Add a source image",
-  "addSourceBody": "Drop, paste, upload, or choose from your Library.",
-  "upload": "Upload",
-  "uploading": "Uploading...",
-  "formats": "PNG / JPG / WEBP",
-  "cameraControls": "Camera controls",
-  "rotationRange": "Rotation (0-360)",
-  "tiltRange": "Tilt (-30 to +30)",
-  "zoomRange": "Zoom (0 to 10)",
-  "degreeUnit": "deg",
-  "estimatedCost": "Estimated cost per run",
-  "generate": "Generate",
-  "generateLocked": "Sign in to generate",
-  "generating": "Generating...",
-  "output": "Output",
-  "outputTitle": "First frame preview",
-  "generatingOverlayTitle": "Generating new angles...",
-  "generatingOverlayBody": "Your last output stays visible until the new run is ready.",
-  "latency": "Latency",
-  "safeGuardrails": "Cinema-safe guardrails were applied to this run.",
-  "emptyTitle": "No output yet",
-  "emptyBody": "Upload an image, tune rotation/tilt/zoom, and click Generate.",
-  "previousJobs": "Previous jobs",
-  "openJobs": "Open Jobs",
-  "angleJob": "Angle job",
-  "close": "Close",
-  "libraryChoose": "Choose from Library",
-  "libraryBody": "Use an uploaded or saved image as the source frame.",
-  "libraryError": "Failed to load library images",
-  "libraryEmpty": "No images saved yet.",
-  "useImage": "Use image",
-  "assetFallback": "Asset",
-  "tabs": {
-    "all": "All",
-    "upload": "Uploaded",
-    "generated": "Generated"
-  },
-  "engineHelpFlux": "FLUX supports eye-level to high-angle only. Negative tilt switches to Qwen automatically for low-angle shots.",
-  "engineHelpQwen": "Low-angle shot detected. Qwen Multiple Angles is now active.",
-  "orbitTitle": "3D Angle Picker",
-  "orbitHint": "Hold and drag to change camera angle",
-  "rotationShort": "Rotation",
-  "tiltShort": "Tilt",
-  "zoomShort": "Zoom",
-  "bestAngles": "Generate from 4 best angles",
-  "addToLibrary": "Add to Library",
-  "alreadyInLibrary": "Already in Library",
-  "download": "Download",
-  "angleLabel": "Angle {index}",
-  "angleSelected": "selected",
-  "priceError": "Unable to load tool pricing",
-  "uploadFailed": "Upload failed",
-  "uploadTooLarge": "Image exceeds {maxMB} MB. Compress it or choose a smaller file.",
-  "invalidUrl": "Paste or drop a valid image URL.",
-  "referenceName": "Reference",
-  "libraryImageName": "Library image",
-  "missingSource": "Upload a source image first.",
-  "generationFailed": "Generation failed",
-  "angleOutputLabel": "Angle Output",
-  "librarySaveFailed": "Library save failed",
-  "authGate": {
-    "title": "Create an account to generate angles",
-    "body": "Guests can explore the tool and browse public examples, but generation, uploads, and Library actions require an account.",
-    "primary": "Create account",
-    "secondary": "Sign in",
-    "close": "Close"
-  },
-  "engines": {
-    "flux-multiple-angles": {
-      "label": "FLUX Multiple Angles",
-      "description": "FLUX LoRA multi-angle engine with horizontal/vertical/zoom controls."
-    },
-    "qwen-multiple-angles": {
-      "label": "Qwen Multiple Angles",
-      "description": "Qwen image edit multiple-angle engine tuned for angle consistency."
-    }
-  }
-} as const;
-
-type AngleCopy = typeof DEFAULT_ANGLE_COPY;
-
-
-const ENGINES = listAngleToolEngines();
-const DEFAULT_ENGINE_ID = ENGINES[0]?.id ?? 'flux-multiple-angles';
-const DEFAULT_UPLOAD_LIMIT_MB = Number.isFinite(Number(process.env.NEXT_PUBLIC_ASSET_MAX_IMAGE_MB ?? '25'))
-  ? Number(process.env.NEXT_PUBLIC_ASSET_MAX_IMAGE_MB ?? '25')
-  : 25;
-const ANGLE_GUEST_EXAMPLE_SOURCE_URL =
-  'https://media.maxvideoai.com/rendersthumbs/301cc489-d689-477f-94c4-0b051deda0bc/d49ec543-8b71-42bb-aa7e-ce5289e28187.webp';
-const ANGLE_GUEST_EXAMPLE_OUTPUT_URL =
-  'https://media.maxvideoai.com/rendersthumbs/301cc489-d689-477f-94c4-0b051deda0bc/44d08767-2bba-4ece-9e37-00991db207af.webp';
-const ANGLE_TOOL_STORAGE_KEY = 'maxvideoai.tools.angle.v1';
-type PersistedAngleToolState = {
-  version: 1;
-  engineId: AngleToolEngineId;
-  params: AngleToolNumericParams;
-  safeMode: boolean;
-  generateBestAngles: boolean;
-  sourceImage: UploadedImage | null;
-  result: AngleToolResponse | null;
-  selectedOutputIndex: number;
-};
-
-function getAngleBillingProductKey(engineId: AngleToolEngineId, generateBestAngles: boolean): string {
-  const family = engineId === 'qwen-multiple-angles' ? 'qwen' : 'flux';
-  return `angle-${family}-${generateBestAngles ? 'multi' : 'single'}`;
-}
-
-function clamp(value: number, min: number, max: number): number {
-  return Math.min(max, Math.max(min, value));
-}
-
-function normalizeRotation(value: number): number {
-  return ((value % 360) + 360) % 360;
-}
-
-function snap(value: number, step: number): number {
-  return Number((Math.round(value / step) * step).toFixed(step < 1 ? 1 : 0));
-}
-
-function sanitizeParams(params: AngleToolNumericParams): AngleToolNumericParams {
-  return {
-    rotation: normalizeRotation(snap(params.rotation, 1)),
-    tilt: clamp(snap(params.tilt, 1), -30, 30),
-    zoom: clamp(snap(params.zoom, 0.1), 0, 10),
-  };
-}
-
-function emitClientMetric(event: string, payload?: Record<string, unknown>) {
-  dispatchAnalyticsEvent(event, payload);
-}
-
-function isAuthRequiredError(error: unknown): boolean {
-  if (!error || typeof error !== 'object') return false;
-  const record = error as { status?: number; code?: string };
-  return record.status === 401 || record.code === 'auth_required' || record.code === 'UNAUTHORIZED';
-}
-
-function formatUsdCompact(value: number | null | undefined): string {
-  if (typeof value !== 'number' || !Number.isFinite(value)) return 'n/a';
-  return `$${value.toFixed(2)}`;
-}
-
-function formatTemplate(template: string, values: Record<string, string | number>): string {
-  return Object.entries(values).reduce((result, [key, value]) => {
-    return result.replaceAll(`{${key}}`, String(value));
-  }, template);
-}
-
-function cleanupSourcePreview(image: UploadedImage | null) {
-  if (!image?.previewUrl?.startsWith('blob:')) return;
-  try {
-    URL.revokeObjectURL(image.previewUrl);
-  } catch {
-    // Ignore preview cleanup failures.
-  }
-}
-
-function collectAnglePreviewImages(
-  renderIds?: string[] | null,
-  renderThumbUrls?: string[] | null,
-  fallbackUrl?: string | null
-): AnglePreviewImage[] {
-  const full = Array.isArray(renderIds) ? renderIds.filter((value): value is string => typeof value === 'string' && value.length > 0) : [];
-  const thumbs = Array.isArray(renderThumbUrls)
-    ? renderThumbUrls.filter((value): value is string => typeof value === 'string' && value.length > 0)
-    : [];
-  const count = Math.max(full.length, thumbs.length);
-
-  if (count > 0) {
-    const outputs: AnglePreviewImage[] = [];
-    for (let index = 0; index < count; index += 1) {
-      const url = full[index] ?? thumbs[index];
-      if (!url) continue;
-      outputs.push({
-        url,
-        thumbUrl: thumbs[index] ?? full[index] ?? null,
-      });
-    }
-    if (outputs.length) return outputs;
-  }
-
-  if (fallbackUrl) {
-    return [{ url: fallbackUrl, thumbUrl: fallbackUrl }];
-  }
-
-  return [];
-}
 
 function AngleOutputMosaic({
   outputs,
@@ -431,116 +201,6 @@ function AngleOutputMosaic({
       })}
     </div>
   );
-}
-
-function parsePersistedAngleToolState(value: string): PersistedAngleToolState | null {
-  try {
-    const raw = JSON.parse(value) as Partial<PersistedAngleToolState> | null;
-    if (!raw || typeof raw !== 'object' || raw.version !== 1) return null;
-    const engineId =
-      raw.engineId === 'flux-multiple-angles' || raw.engineId === 'qwen-multiple-angles'
-        ? raw.engineId
-        : null;
-    if (!engineId) return null;
-    const params =
-      raw.params && typeof raw.params === 'object'
-        ? sanitizeParams({
-            rotation: Number((raw.params as AngleToolNumericParams).rotation ?? 0),
-            tilt: Number((raw.params as AngleToolNumericParams).tilt ?? 0),
-            zoom: Number((raw.params as AngleToolNumericParams).zoom ?? 0),
-          })
-        : { rotation: 8, tilt: 2, zoom: 1.2 };
-    const sourceImage =
-      raw.sourceImage &&
-      typeof raw.sourceImage === 'object' &&
-      typeof raw.sourceImage.url === 'string' &&
-      raw.sourceImage.url.trim().length
-        ? {
-            id: typeof raw.sourceImage.id === 'string' ? raw.sourceImage.id : null,
-            url: raw.sourceImage.url,
-            previewUrl:
-              typeof raw.sourceImage.previewUrl === 'string' && raw.sourceImage.previewUrl.trim().length
-                ? raw.sourceImage.previewUrl
-                : raw.sourceImage.url,
-            width: typeof raw.sourceImage.width === 'number' ? raw.sourceImage.width : null,
-            height: typeof raw.sourceImage.height === 'number' ? raw.sourceImage.height : null,
-            name: typeof raw.sourceImage.name === 'string' ? raw.sourceImage.name : null,
-            source:
-              raw.sourceImage.source === 'upload' || raw.sourceImage.source === 'library' || raw.sourceImage.source === 'paste'
-                ? raw.sourceImage.source
-                : undefined,
-          }
-        : null;
-    const result =
-      raw.result &&
-      typeof raw.result === 'object' &&
-      Array.isArray(raw.result.outputs)
-        ? (raw.result as AngleToolResponse)
-        : null;
-    const selectedOutputIndex =
-      typeof raw.selectedOutputIndex === 'number' && Number.isFinite(raw.selectedOutputIndex)
-        ? Math.max(0, Math.floor(raw.selectedOutputIndex))
-        : 0;
-
-    return {
-      version: 1,
-      engineId,
-      params,
-      safeMode: raw.safeMode !== false,
-      generateBestAngles: raw.generateBestAngles === true,
-      sourceImage,
-      result,
-      selectedOutputIndex,
-    };
-  } catch {
-    return null;
-  }
-}
-
-function getUploadTooLargeMessage(copy: AngleCopy, maxMB: number): string {
-  return formatTemplate(copy.uploadTooLarge, { maxMB });
-}
-
-async function uploadImage(file: File, copy: AngleCopy): Promise<UploadedImage> {
-  const preparedFile = await prepareImageFileForUpload(file, {
-    maxBytes: DEFAULT_UPLOAD_LIMIT_MB * 1024 * 1024,
-  });
-
-  const formData = new FormData();
-  formData.set('file', preparedFile, preparedFile.name);
-
-  const response = await authFetch('/api/uploads/image', {
-    method: 'POST',
-    body: formData,
-  });
-
-  const payload = (await response.json().catch(() => null)) as
-    | {
-        ok?: boolean;
-        error?: string;
-        maxMB?: number;
-        asset?: {
-          url: string;
-          width?: number | null;
-          height?: number | null;
-          name?: string | null;
-        };
-      }
-    | null;
-
-  if (!response.ok || !payload?.ok || !payload.asset?.url) {
-    if (payload?.error === 'FILE_TOO_LARGE' || response.status === 413) {
-      throw new Error(getUploadTooLargeMessage(copy, payload?.maxMB ?? DEFAULT_UPLOAD_LIMIT_MB));
-    }
-    throw new Error(payload?.error ?? `Upload failed (${response.status})`);
-  }
-
-  return {
-    url: payload.asset.url,
-    width: payload.asset.width,
-    height: payload.asset.height,
-    name: payload.asset.name,
-  };
 }
 
 function AngleOrbitSelector({
