+++ frontend/src/components/tools/angle/_lib/angle-workspace-copy.ts
@@ -1,0 +1,96 @@
+export const DEFAULT_ANGLE_COPY = {
+  "disabledTitle": "Tools are disabled",
+  "disabledBody": "Enable `FEATURES.workflows.toolsSection` to access this area.",
+  "back": "Back to Tools",
+  "engine": "Engine",
+  "engineSelect": "Engine selection",
+  "sourceImage": "Source image",
+  "sourceReady": "Source image ready",
+  "sourceFromLibrary": "From Library",
+  "sourceFromPaste": "From pasted URL",
+  "sourceFromDevice": "Uploaded from device",
+  "sourceFromExample": "Public example",
+  "replace": "Replace",
+  "library": "Library",
+  "remove": "Remove",
+  "sourceAlt": "Source",
+  "addSourceTitle": "Add a source image",
+  "addSourceBody": "Drop, paste, upload, or choose from your Library.",
+  "upload": "Upload",
+  "uploading": "Uploading...",
+  "formats": "PNG / JPG / WEBP",
+  "cameraControls": "Camera controls",
+  "rotationRange": "Rotation (0-360)",
+  "tiltRange": "Tilt (-30 to +30)",
+  "zoomRange": "Zoom (0 to 10)",
+  "degreeUnit": "deg",
+  "estimatedCost": "Estimated cost per run",
+  "generate": "Generate",
+  "generateLocked": "Sign in to generate",
+  "generating": "Generating...",
+  "output": "Output",
+  "outputTitle": "First frame preview",
+  "generatingOverlayTitle": "Generating new angles...",
+  "generatingOverlayBody": "Your last output stays visible until the new run is ready.",
+  "latency": "Latency",
+  "safeGuardrails": "Cinema-safe guardrails were applied to this run.",
+  "emptyTitle": "No output yet",
+  "emptyBody": "Upload an image, tune rotation/tilt/zoom, and click Generate.",
+  "previousJobs": "Previous jobs",
+  "openJobs": "Open Jobs",
+  "angleJob": "Angle job",
+  "close": "Close",
+  "libraryChoose": "Choose from Library",
+  "libraryBody": "Use an uploaded or saved image as the source frame.",
+  "libraryError": "Failed to load library images",
+  "libraryEmpty": "No images saved yet.",
+  "useImage": "Use image",
+  "assetFallback": "Asset",
+  "tabs": {
+    "all": "All",
+    "upload": "Uploaded",
+    "generated": "Generated"
+  },
+  "engineHelpFlux": "FLUX supports eye-level to high-angle only. Negative tilt switches to Qwen automatically for low-angle shots.",
+  "engineHelpQwen": "Low-angle shot detected. Qwen Multiple Angles is now active.",
+  "orbitTitle": "3D Angle Picker",
+  "orbitHint": "Hold and drag to change camera angle",
+  "rotationShort": "Rotation",
+  "tiltShort": "Tilt",
+  "zoomShort": "Zoom",
+  "bestAngles": "Generate from 4 best angles",
+  "addToLibrary": "Add to Library",
+  "alreadyInLibrary": "Already in Library",
+  "download": "Download",
+  "angleLabel": "Angle {index}",
+  "angleSelected": "selected",
+  "priceError": "Unable to load tool pricing",
+  "uploadFailed": "Upload failed",
+  "uploadTooLarge": "Image exceeds {maxMB} MB. Compress it or choose a smaller file.",
+  "invalidUrl": "Paste or drop a valid image URL.",
+  "referenceName": "Reference",
+  "libraryImageName": "Library image",
+  "missingSource": "Upload a source image first.",
+  "generationFailed": "Generation failed",
+  "angleOutputLabel": "Angle Output",
+  "librarySaveFailed": "Library save failed",
+  "authGate": {
+    "title": "Create an account to generate angles",
+    "body": "Guests can explore the tool and browse public examples, but generation, uploads, and Library actions require an account.",
+    "primary": "Create account",
+    "secondary": "Sign in",
+    "close": "Close"
+  },
+  "engines": {
+    "flux-multiple-angles": {
+      "label": "FLUX Multiple Angles",
+      "description": "FLUX LoRA multi-angle engine with horizontal/vertical/zoom controls."
+    },
+    "qwen-multiple-angles": {
+      "label": "Qwen Multiple Angles",
+      "description": "Qwen image edit multiple-angle engine tuned for angle consistency."
+    }
+  }
+} as const;
+
+export type AngleCopy = typeof DEFAULT_ANGLE_COPY;
